vmware: Fix double restartPriority override
Adding an override for restartPriority when one alrady exists is an
error in the VMware API. We add the override twice when running
`finish_migration()`: inside `_resize_vm()` and again explicitly at
the end. Remove the one in `_resize_vm()`.

Additionally, remove the *removal* as well -- under the assumption
that every resize creates a new VM object with a new moref, so
explicit deletion never becomes necessary.

Fixes
    f157f63707 'vmware: Restart BigVMs with "high" priority'

Change-Id: Ifdec94dcdedfcb8bf143ec5d5435058c8a94cffe