quota:separate: Fix os-quota-class-sets create endpoint
The endpoint was not checking the context against the policy, so every
valid token was able to create custom quotas. Additionally, it wasn't
checking the quota_class_name string length and it JSON input was not
validated against any schema.

We update the "create()" method to contain similar code to the
"_update()" method, adding a policy for it.

We also validate the schema now and thus have to update the schema to
allow our custom "instances_"-prefixed quotas.

Change-Id: I78b21227f800434029ac753464a8c3feee1e236f