Disconnecting volume from the compute host
cmd nova-manage volume_attachment refresh vm-id vol-id connetor

There were cases where the instance said to live in compute#1 but the
connection_info in the BDM record was for compute#2, and when the script
called `remove_volume_connection` then nova would call os-brick on
compute#1 (the wrong node) and try to detach it.

In some case os-brick would mistakenly think that the volume was
attached (because the target and lun matched an existing volume on the
host) and would try to disconnect, resulting in errors on the compute
logs.

- Added HostConflict exception
- Fixes dedent in cmd/manange.py
- Updates nova-manage doc

Conflicts: stable/2023.2
       - nova/exception.py
       changes:
       (I052441076c677c0fe76a8d9421af70b0ffa1d400)
       Removed 2 new EphemeralEncryption exceptions as feature is not backportable.

Closes-Bug: #2012365
Change-Id: I21109752ff1c56d3cefa58fcd36c68bf468e0a73
(cherry picked from commit a8f81d5f08692c16d538af3197460d9426cc00a1)