Add compute_id columns to instances, migrations
This adds a compute_id column to the instances table in anticipation
of later patches to link instances to computes by row id instead of
name. It also adds a dest_compute_id to migrations, so that we can
positively track where the instance is to end up.

Related to blueprint compute-object-ids

Change-Id: Ia9c367c08604926598b3c1c21bd8dbd2ba0bfbc6