Remove focal job for 2023.2
Neutron also flipped to python>=3.9 on all their repos this morningi[1]
which means we can't install neutron on focal at all. I'm not sure if
that's going to get reverted at this point, but even if it is, it's
going to take a while to undo. As noted in the comments and the
original commit[2] that added this job, it was intended to be removed
when we dropped focal from the test interface, which we have now done.

1: https://review.opendev.org/q/topic:bug%252F2017478
2: https://review.opendev.org/c/openstack/nova/+/861111

Change-Id: I5be638a702629e07ec9c88bd67bb9b7f1212f7fc