testing: Use inspect.isfunction() to check signatures
We recently discovered that nova.test.SubclassSignatureTestCase has not
worked properly since support for python 2.7 was dropped.

The reason is the use of the inspect.ismethod() function; in python 2.7
it returned true if the object was a bound or an unbound method [2] but
as of python 3, it only returns true if the object is a bound method
[2].

Because of this, LibvirtBaseVolumeDriverSubclassSignatureTestCase was
not actually checking any signatures.

This replaces the use of inspect.ismethod() with inspect.isfunction()
to fix the issue.

It also adds a skip for subclass private functions as these are
expected to be different among subclasses.

One subclass method signature caught by the test is also updated to
match the base class.

[1] https://docs.python.org/2.7/library/inspect.html#inspect.ismethod
[2] https://docs.python.org/3.5/library/inspect.html#inspect.ismethod

Closes-Bug: #2019772

Change-Id: I5644effac7206035de77e0d1cd450a9fb96c3a0d