Add debug logging when Instance raises OrphanedObjectError
This logging would be helpful in debugging issues when
OrphanedObjectError is raised by an instance. Currently, there is
not a way to identify which instance is attempting to lazy-load a
field while orphaned. Being able to locate the instance in the
database could also help with recovery/cleanup when a problematic
record is disrupting operation of a deployment.

Change-Id: I093de2839c1bb7c949a0812e07b63de4cc5ed167
(cherry picked from commit e0fbb6fc06d3b08b938af2e36b11f04c57fe6954)