Call Neutron immediately upon _post_live_migration() start
Previously, we first called to Cinder to cleanp source volume
connections, then to Neutron to activate destination port bindings.
This means that, if Cinder or the storage backend were slow, the live
migrated instance would be without network connectivity during that
entire process.

This patch immediately activates the port bindings on the destination
(and thus gives the instance network connectivity). We just need to
get the old network_info first, in order to use it in notifications
and to stash it for the later call to
driver.post_live_migration_at_source().

This is a smaller and safer change than the parallelization attempt in
the subsequent patch, so it's done in its own patch because it might
be backportable, and would help with network downtime during live
migration.

To avoid any potential data leaks, we want to be certain to cleanup
volumes on the source. To that end we wrap the code that is being
moved before the source volume cleanup code in a try/finally block
in order to prevent any uncaught exception from blocking the cleanup.

Change-Id: I700943723a32e732e3e3be825f3fd44a9f923a0b
(cherry picked from commit 26fbc9e8e7d353e66739f910865d0b6498811bb0)