vmware: Fix rescue for efi-booted VMs
VMware by default enables a "quickBoot" feature for efi-booted VMs,
which exposes only the first disk of the VM when finding out which disk
to boot from. This leads to the VM not booting the rescue disk and thus
ignoring the explicitly set "bootOptions.bootOrder" property.

There's an advanced setting called "efi.quickBoot.enabled" to disable
the default behavior by providing a value of "FALSE". We set this
property on rescue and unset it on unrescue, as it might have an adverse
effect on the boot time when the underlying disks are slow.

The device order persists, and may refer to another device.  That breaks
cloning the VM for the purpose of a snapshot, where the disk gets
removed.

Co-Authored-By: Fabian Wiesel <fabian.wiesel@sap.com>
Change-Id: I64cf1cd4478a9c48a4cce723d224dac31edf85c6
Incorporated-Change-Id: I90d7890f6b10622b7a6efc8e988d7736a3e87df2