compute: Do not re-schedule on quota error
When creating an instance without having enough Port quota, the cause of
the failure can be hidden from the user. This happens, if the number
hosts found as candidates for the instance are less than the number of
scheduling attempts Nova undertakes (by default 3). The reason for this
is, that `nova-conductor` encounters no more hosts instead of
`nova.scheduler.utils.populate_retry()` encounters the max number of
retries.

Nova re-schedules the instance's build, because the `network_info`
object is built asynchronously. This means, the quota problem only comes
up when first accessing the object's data - which happens during disk
prep for Ironic and during `spawn()` for VMware. We handle the
`QuotaError` and `OverQuota` exceptions during disk prep already, but
the call to `spawn()` doesn't. Therefore, these exceptions go into the
default branch handling `Exception`, which re-schedules the instance.

To fix this, we add `QuotaError` and `OverQuota` to the list of handled
exceptions raising a `BuildAbortException` with the exception's message.

NOTE: When forward-porting this commit to after
I17a3e20b8be98f9fb1a04b91fcf1237d67165871, `QuotaError` doesn't exist
anymore as it became the same as `OverQuota`.

Change-Id: I2a2a95d23defcec3db8b5045a59c3852a4a5785b