Enforce correct choice for [DEFAULT] enabled_ssl_apis
The [DEFAULT] enabled_ssl_apis option is supposed to contain the subset
(or the same set) of the [DEFULT] enabled_apis option.

Because we accept only the correct values for enabled_apis, we can do
the same for enabled_ssl_apis. This would allow users to notice any
wrong values which were ignored previously.

Change-Id: Ibba8f3ea307df3e81e18c3a75361e27a0f52701d