vmware: Add CustomField helper class in vim_util
This class bundles together the helper functions around CustomFieldDef
and CustomFieldsManager so they are clearly visible. We could have also
used a module for the same purpose, but opted for `vim_util` to not
create more files and to not have to split `vim_util` into a directoy
with multiple files and thus stray more from upstream.

The helper class takes a session on init, so handling multiple calls
makes for shorter argument lists and we assume there will be usually
more than one call into the instance.

In the end, this functionality could move into `oslo.vmware`.

Change-Id: Iea02d23b346e35efe6d1a2487c32d7910b72ffae