vmware: Add sync of traits to cluster custom attributes
To help operations colleagues, we now sync the CUSTOM_* traits and the
COMPUTE_STATUS_DISABLED trait into the custom attributes of a cluster.
This helps their automation for configuring cluster and also helps during
trouble shooting session, as operations colleagues don't necessarily
have access to Placement and Nova admin APIs.

To implement this, we add a sync-loop running by default every 10min.
This loop will create/delete CustomFieldDef objects via the
`CustomField` helper class and set values to "true" according to the
compute host's traits.

We use "openstack.nova.traits." as prefix for the custom attributes. We
keep the full trait name and do not remove the `CUSTOM_` prefix, to make
it easier to find the traits in Placement and because we also sync some
non "CUSTOM_"-prefixed traits.

Change-Id: I7df8d3b97e086f304db38e25b62eb1725ffd1d3a