Log excessive lazy-loading behavior
This makes us keep track of how many things we've lazy-loaded on a
given object instance in a single service. Ideally we'd only ever
lazy-load something zero or one times where it makes sense for
easier optimization. However, because this 'works so well' we end up
lazy-loading more things than we expect.

Change-Id: Ie379239c5de795ba8acc281da9208b3a10e92da0