[functional] Don't leak running live migration
The
test_get_migrations_after_live_migrate_server_in_different_project test
case took 160 sec to finish. It turned out that it start a live
migration and the never cancels it, this result in a leaked bare
greenlet. At the end of the test case the compute service is stopped and
we introduced a wait for all running greenlets at that point in
Ia916310522b007061660172fa4d63d0fde9a55ac. This wait take 160 sec to
time out.

So this change deletes the instance before the end of the test case to
cancel the live migration and avoid leaking a greenlet and make the test
case fast again.

Change-Id: I8341e328d4f5da5edc069eb69988fccb07353839