[functional] Avoid leaking greenlet in UnifiedLimits tests
The leak tests just start a server create and the finish so I think
the actual conductor thread leaks. If I add a bit of sleep to the end of
these test cases then the leak disappears. Unfortunately adding the
sleep at the greenpool fixture does not have this effect. So instead I
added delete_server calls to the end of these tests as those are nicer
than sleeps and has the same effect regarding the leak.

We are down to 3 leaking functional test

Change-Id: I070390c695283bdd9b87cd879aa2a9257ee7bdfb