[functional]Fix remaining greenlet leaks
We leak due to running background operations like server create, rebuild
and in one case a sleeping claim. So this patch removes the leaks by
making sure that the background operations stop before the test
finishes.

As there is no more leak in functional test too this patch makes the
leak an error there too.

Change-Id: I6905999050e8d09b772837034a212c534e9c3226