disable ovn based testing of test_live_migration_with_trunk
due to bug #1940425 where ml2/ovn is not correctly configuring the
active status on trunk port we see test_live_migration_with_trunk
fail more often then not.

This was previously disbaled in tempest and then fixed in neutorn.
the tempest skip was then reverted and so was the neutron fix as
it broke somethign else... so this is failing in our gate again.

This change skips test_live_migration_with_trunk on all jobs that
are using ml2/ovn but keeps it enabled on the hybrid plug job
which uses ml2/ovs.

Related-Bug: #1940425
Change-Id: I0a8dd6e6e30526aa2841b4db67ed9affed166fd8