Sanity check that new hosts have no instances
If we are starting up and we think we are a new host (i.e. no pre-
existing service record was found for us), we expect to have no
instances on our hypervisor. If that is not the case, it is likely
that we got pointed at a new fresh database or the wrong database
(i.e. a different cell from our own). In that case, we should abort
startup to avoid creating new service, compute node, and resource
provider records.

This is a sort of follow-on additional improvement related to
work done in blueprint stable-compute-uuid.

Change-Id: Id817c51c90485119270f3b7f3c52858f42100637