Do not manage CPU0's state
If cpu_power_management_strategy is "cpu_state" and CPU0 is in the
dedicated set, we should just ignore it whenever we go to manage the
state. Since CPU0 cannot be powered off, but may be otherwise suitable
for the dedicated set, we can just skip it whenever we would normally
go to power it up or down.

Change-Id: I995c0953b361c7016bd77482fa2e2f276d239828
Fixes-Bug: #2038840