Expose volume sizes to scheduler during resize/migration
We want to allow customers to prevent resizes between shard during
resizes, if the sum of the volume sizes hit a certain, Nova-configured
threshold. Since we only need volume sizes for that, we just provide
volumes sizes in the scheduler hint.

The information we store in the "volume_sizes" scheduler-hint can change
at any time so is not worth storing in the DB after the scheduling is
done.  Therefore, we remove it again once the resize got triggered and
before saving the `RequestSpec`.

Change-Id: If1ba024dd2337f26a38148179b16a7cfd82de64c