scheduler: Move project tag caching to mixin
We have an implementation caching certain tags for projects in the
`ShardFilter`. In the future, we want to add another filter to the
scheduler that will also react to project tags. Therefore, we move the
code handling the caching out into a Mixin class.

The Mixin supports prefix and exact matching, configured via class
attribute.

Change-Id: I3ad865ec4dbbf097fc4718b0753b86a8e64775fa