scheduler: Add VolumeSizeSameShardResizeFilter
We want to allow customers to optionally disable resizes of VMs across
shards if the sum of the volume sizes reach a certain threshold. With
this, we want customers to have more reliable resize times that might
fit in their downtime windows as resizing across shards can create
volume migrations which can take a long time.

The threshold is not configurable by the user and set to 1 TiB by
default.  It can be changed with the
`volume_size_same_shard_resize_filter_sum` config setting.

Change-Id: I13a0e16713b90954cb48ea7f0a7c6db870d60bdc