vmware: Add CPU reservation support
To reserve CPUs a new extra-spec(-section) is introduced:
    reservation:cpu

This is expected to be set to equal the number of total CPUs of the
flavor, just like the CUSTOM_MEMORY_RESERVABLE resource.

The reason why this is not a custom resource, is that we don't need
the additional complexity for CPUs, because this is only a performance
setting according to VMware. We are not running non-cpu-reserved
flavored VMs alongside ones with reserved CPUs, instead only using
this on clusters where _all_ VMs get their CPUs reserved.

CPU reservation in VMware is set in MHz, not in number-of-vCPUs. This
necessitates getting the median CPU frequency of the hosts from the
cluster, before reservation can be calculated and set.

VMware reports CPU resources in MHz, but this is in relation to the
physical CPU count, not the hyperthreaded CPU count. Check for the
pCPU-to-vCPU ratio (which is either 1 = "no hyperthreading",
or 2 = "hyperthreaded") and adjust the per-CPU MHz number of the host.

The code is very defensive: The first division is inverted as to avoid
a second division in the other line. If either `threads` or `pcpus` is
0 then 1 will be substituted so that the ratio will be 1 in the case
of missing or zero stats.

Add test cases for `get_reserved_memory_and_cpu()` in a new
`VmSpecsTestCase`, named so for in the future adding tests for other
functions from that part of nova.utils.py.

Alternatively it would have been possible to re-use:
    quota:cpu_reservation
We don't do that because it doesn't take into account the cpu-to-mhz
conversion and the "quota:" prefix doesn't really fit.

Follow-up-for: Iad0e9bf421f1c4fd793a4e738f105b3c06e4c045
Change-Id: I907278b970d5ac658223e821b46461eeac2a5275
Incorporated-Change-Id: I10df0902ea6ff27d1d32d20bf6c6d8c597184e22
Incorporated-Change-Id: Ic06bb756d37e2d552c1b734b8600a6cdc08b3c08
Incorporated-Change-Id: I5cb406e53b32b71a802a178b5b7c635657779822