Add custom API to vMotion VM inside cluster
Since the vmwareapi driver manages a whole cluster, it's possible to
move VMs on the same Nova host. This is not a use-case other drivers can
support. Therefore, we add it as a custom API endpoint.

We want to support this, because services in need to vMotion don't
necessarily have access to the vCenter or the time to implement vCenter
APIs.

Change-Id: I4069fc4fe1fad94604f20702f100a49eea06fa44