[ironic] Partition & use cache for list_instance*
list_instances and list_instance_uuids, as written in the Ironic driver,
do not currently respect conductor_group paritioning. Given a nova
compute is intended to limit it's scope of work to the conductor group
it is configured to work with; this is a bug.

Additionally, this should be a significant performance boost for a
couple of reasons; firstly, instead of calling the Ironic API and
getting all nodes, instead of the subset (when using conductor group),
we're now properly getting the subset of nodes -- this is the optimized
path in the Ironic DB and API code. Secondly, we're now using the
driver's node cache to respond to these requests. Since list_instances
and list_instance_uuids is used by periodic tasks, these operating with
data that may be slightly stale should have minimal impact compared to
the performance benefits.

Conflicts:
  nova/tests/unit/virt/ironic/test_driver.py
  nova/virt/ironic/driver.py

NOTE(elod.illes): conflict is caused by openstacksdk related feature
patch (I9fb5a385487f601329bccc9efde325d86be397b6) which was merged in
development branch (2024.1 Caracal), hence it does not exist in 2023.2.

Closes-bug: #2043036
Change-Id: If31158e3269e5e06848c29294fdaa147beedb5a5
(cherry picked from commit fa3cf7d50cba921ea67eb161e6a199067ea62deb)